Added Login and Logout Servlets
1) Why is the change being made?
---------------------------------
Users will eventually be able to log in and log out of the website and store information about themselves.
This change is the first step in that process.

2) What has changed?
---------------------------------
The Login and Logout Servlet classes have been created which will log the user in or out respectivly on GET request.

3) Other concerns and/or dependent changes?
---------------------------------
Currently UserService is being used to log in users, in the future we should change this to OAuth.