Add UserInformationServlet with CRUD functionality
Description:
UserInformationServlet is responsible for Creation, Reading, Updating,
and Deleting User Objects. The User Objects are stored with DataStore,
and contain information about users who log on using UserService.

Changes:
UserInformationServlet is implimented and active.
index.html now has a post field for interacting with
UserInformationServlet.

Purpose:
This allows our app to have control of what is being stored indepndent
of the UserService. Currently its main purpose is to hold user
displayNames, in the future it could tie Users to profilePictures.

Note:
Creation, Updates, and Deletion are only allowed if the User being acted
on is logged in.

Security Concerns:
1) Anyone can Read anyone's data so long as they know their user-id
Not a huge concern, UserService isn't intended to hold sensitive
information.