Add Login/Logout/Account functionality through UI
1) Why is the change being made?
---------------------------------
The Users will need to be able to handle logging in and creation updating and deletion of accounts,
and for this to be possible it should be allowed via the UI.

2) What has changed?
---------------------------------
Login, Logout, Account Creation, and Account Updating functionalty has all been added to the UI.

3) Other concerns and/or dependent changes?
---------------------------------
We still need to add account deletion, and add user accessable account reading. Currently view account shows the
datastore json object for the current user. We can use this to make a page to show their current account status.
Ideally this page would have similarity to the edit_account page.