Add PartyPlayer and Parties utility class
1) Why is the change being made?
---------------------------------
Each Party room will be able to play music in sync, we are accomplishing this by having all users connected to a party listen to
the server to know exactly where in a song it should be at any time. However before this change, the Server didn't store what songs would be playing,
and it didn't record where in a song it would be at any given point in time.

This change introduces the PartySongPlayer which is a Lazy implimentation of a typical music player. Instead of updating the playlist after a song completes,
this player updates itself whenever someone accesses it, this is possible because it can knows how much time has passed since the last access, and can therefore
quickly simulate all the actions that occured while nobody accessed it. This also adds quite a bit of functionality to the Party Rooms so the Parties utility class
has also been made to make interacting with data tied to parties a bit easier.

Additionally the YoutubeSongPlayInfo data class has been added to be searilize information about the current song playing in a Party at any given time.

2) What has changed?
---------------------------------
The PartySongPlayer and YoutubeSongPlayInfo classes have been implemented, and the PartyMusicPlayerServlet allows RESTful access to them.
Now using a HTTP GET request one can see the current song play info for a given room, and with a HTTP POST method one can start/stop the player, seek to a time in the song,
add songs to be played, or skip to a current time in the song.

They also both make good use of the Parties utility class which has also been implimented and allows easy access to party information and the players attached to them.

3) Testing
---------------------------------
Heavy use of PostMan to test post requests were properly taking place. From using it a bit it seems like all the changes are working as intended.