Add Playlist to party.html and java methods to enable that
1) Why is the change being made?
---------------------------------
Users would want to see the songs that will be played to more easily communicate what the site is doing at any given time.

2) What has changed?
---------------------------------
Fist the PartyPlaylistState java class has been created and this is now what the PartyMusicPlayerServlet returns on a GET request.
The PartySongPlayer also has a new method added to it to return the current playlist, which is one of the components of the PartyPlaylistState object.
And the PartyMusicPlayerServlet now is set up properly to return the PartyPlaylistState object.
This newly returned object is also set up to put the playlist into a playlist-container div now added in the party.html file on a 2000ms refresh.

3) Other concerns and/or dependent changes?
---------------------------------
The playlist-container currently is configured to look identical to the comments-container, this is fine for right now, but ideally the playlist would appear
on the right of the page and possibly have more formatting, for example song images.