Automate username filling for logged in users
1) Why is the change being made?
---------------------------------
If you're logged in you shouldn't have to type in your username.

2) What has changed?
---------------------------------
The username box for commenting is only shown to logged out users, and if you're logged in then
it uses your configured displayname.

3) Other concerns and/or dependent changes?
---------------------------------
If you configure your display name to the empty-string then you can't comment.
This can be fixed by setting your display name. This is a pretty bad problem, we should show users this issue so they can fix it,
but for the alpha this known bug is admissable.