Synchronize Parties utility class Creation method
1) Why is the change being made?
---------------------------------
The Parties utility class used the ConcurrentHashMap to store PartyPlayers, creating multiple objects with a put is dangerous because
the ConcurentHashMap implementation can return stale values. This had the potential to make different calls on the site return different players with different states.

2) What has changed?
---------------------------------
The createOrReplace method has been replaced with a create, and if there is already an existing object its a no-op. This eliminates the creation of stale values making
ConcurrentHashMap still a fine choice for storing PartyPlayers.