Make frontend player start a stopped backend player on song addition
1) Why is the change being made?
---------------------------------
If the backend player is empty it sets its state to stopped, this can cause discrepencies between the frontend and backend.
Before the frontend only started the backend when the page was loaded, this could lead to problems because if a song finishes the backend will stop,
someone will add a song to be played, the frontend will start playing the song, and the backend will remain stopped.

2) What has changed?
---------------------------------
Now if someone adds a song to a stopped player the player will automatically restart to ensure users can listen to newly added songs.

3) Other concerns and/or dependent changes?
---------------------------------
The player is still restarted on page load, in the future this should be changed to not occur. Users should be able to start and stop the player indifferent from page reoloads.